Parent Button to Control Group
Added ability for master toggle to control group of child elements. When the parent is clicked, all the children in the group will change to match the status of the parent. Parent can display 3 states: Off (grey when all children are off), On (green when all children are on), or Partial (yellow when some but not all children are on)